# frontend/public/openapi.yaml
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"tags":[{"name":"Dragon","description":"Операции над драконами"},{"name":"Authentication","description":"Регистрация и аутентификация пользователей"},{"name":"Special","description":"Специальные бизнес-операции"},{"name":"Get","description":"Получение списков сущностей"},{"name":"Create","description":"Создание базовых сущностей"},{"name":"Delete","description":"Удаление сущностей"}],"paths":{"/updateDragonById/{id}":{"put":{"tags":["Dragon"],"summary":"Обновить дракона по ID","description":"Обновляет данные дракона и возвращает DTO.","operationId":"updateDragonById","parameters":[{"name":"id","in":"path","description":"ID дракона","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"requestBody":{"description":"Обновлённые данные дракона","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Dragon"}}},"required":true},"responses":{"200":{"description":"Обновлено","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DragonDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/dragons/kill":{"post":{"tags":["Special"],"summary":"Убить дракона","description":"Отмечает дракона как убитого, связывая с указанным убийцей. 404 — если дракон/убийца не найдены, 409 — если уже убит.","operationId":"killDragon","requestBody":{"description":"ID дракона и убийцы","content":{"application/json":{"schema":{"$ref":"#/components/schemas/KillDragonDTO"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Dragon"}}}},"404":{"description":"Не найдено","content":{"*/*":{"schema":{"type":"string"}}}},"409":{"description":"Конфликт состояния","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/createPerson":{"post":{"tags":["Create"],"summary":"Создать персонажа","description":"Создаёт персонажа. Если указана локация по id — будет привязана к персонажу.","operationId":"createPerson","requestBody":{"description":"Персонаж","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PersonDTO"}}},"required":true},"responses":{"201":{"description":"Создано","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PersonDTO"}}}},"400":{"description":"Ошибка создания","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/createLocation":{"post":{"tags":["Create"],"summary":"Создать локацию","description":"Создаёт локацию и возвращает исходный DTO.","operationId":"createLocation","requestBody":{"description":"Локация","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LocationDTO"}}},"required":true},"responses":{"201":{"description":"Создано","content":{"*/*":{"schema":{"$ref":"#/components/schemas/LocationDTO"}}}},"400":{"description":"Ошибка создания","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/createHead":{"post":{"tags":["Create"],"summary":"Создать голову дракона","description":"Создаёт голову дракона и возвращает исходный DTO.","operationId":"createHead","requestBody":{"description":"Голова дракона","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DragonHeadDTO"}}},"required":true},"responses":{"201":{"description":"Создано","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DragonHeadDTO"}}}},"400":{"description":"Ошибка создания","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/createDragon":{"post":{"tags":["Dragon"],"summary":"Создать дракона","description":"Создаёт дракона на основании составного запроса. Отправляет обновлённый список по WebSocket.","operationId":"createDragon","requestBody":{"description":"Данные для создания дракона","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NewDragonResp"}}},"required":true},"responses":{"201":{"description":"Создано","content":{"*/*":{"schema":{"type":"string"}}}},"500":{"description":"Внутренняя ошибка","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/createCoordinates":{"post":{"tags":["Create"],"summary":"Создать координаты","description":"Создаёт координаты и возвращает исходный DTO.","operationId":"createCoordinates","requestBody":{"description":"Координаты","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CoordinatesDTO"}}},"required":true},"responses":{"201":{"description":"Создано","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CoordinatesDTO"}}}},"400":{"description":"Ошибка создания","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/createCave":{"post":{"tags":["Create"],"summary":"Создать пещеру","description":"Создаёт пещеру и возвращает исходный DTO.","operationId":"createCave","requestBody":{"description":"Пещера дракона","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DragonCaveDTO"}}},"required":true},"responses":{"201":{"description":"Создано","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DragonCaveDTO"}}}},"400":{"description":"Ошибка создания","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/auth/sign-up":{"post":{"tags":["Authentication"],"summary":"Регистрация пользователя","description":"Создаёт нового пользователя и возвращает форму ответа с токеном/данными.","operationId":"authSignUp","requestBody":{"description":"Данные для регистрации","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpRequest"}}},"required":true},"responses":{"200":{"description":"Пользователь создан","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRespForm"}}}},"400":{"description":"Ошибка валидации","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/auth/sign-in":{"post":{"tags":["Authentication"],"summary":"Вход пользователя","description":"Аутентифицирует пользователя и возвращает форму ответа (обычно с JWT).","operationId":"authSignIn","requestBody":{"description":"Данные для входа","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInRequest"}}},"required":true},"responses":{"200":{"description":"Успешная аутентификация","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AuthRespForm"}}}},"401":{"description":"Неверные учётные данные","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/getPerson":{"get":{"tags":["Get"],"summary":"Получить персонажей","operationId":"getPersons","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/PersonDTO"}}}}}},"security":[{"bearerAuth":[]}]}},"/getLocation":{"get":{"tags":["Get"],"summary":"Получить локации","operationId":"getLocations","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LocationDTO"}}}}}},"security":[{"bearerAuth":[]}]}},"/getHead":{"get":{"tags":["Get"],"summary":"Получить головы драконов","operationId":"getHeads","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DragonHeadDTO"}}}}}},"security":[{"bearerAuth":[]}]}},"/getDragons":{"get":{"tags":["Dragon"],"summary":"Получить всех драконов","description":"Возвращает полный список драконов.","operationId":"getDragons","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/Dragon"}}}}}},"security":[{"bearerAuth":[]}]}},"/getDragonById/{id}":{"get":{"tags":["Dragon"],"summary":"Получить дракона по ID","description":"Возвращает дракона по идентификатору.","operationId":"getDragonById","parameters":[{"name":"id","in":"path","description":"ID дракона","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Dragon"}}}},"404":{"description":"Не найден","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Dragon"}}}}},"security":[{"bearerAuth":[]}]}},"/getCoordinates":{"get":{"tags":["Get"],"summary":"Получить координаты","operationId":"getCoordinates","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CoordinatesDTO"}}}}}},"security":[{"bearerAuth":[]}]}},"/getCaves":{"get":{"tags":["Get"],"summary":"Получить пещеры","operationId":"getCaves","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DragonCaveDTO"}}}}}},"security":[{"bearerAuth":[]}]}},"/dragons/oldest":{"get":{"tags":["Special"],"summary":"Самый старый дракон","operationId":"getOldestDragon","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DragonDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/dragons/maxCave":{"get":{"tags":["Special"],"summary":"Дракон с максимальной пещерой","operationId":"getDragonWithMaxCave","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DragonDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/dragons/headGreater/{param}":{"get":{"tags":["Special"],"summary":"Дракон с головой больше параметра","description":"Ищет дракона, у которого размер головы больше указанного параметра.","operationId":"getDragonHeadGreater","parameters":[{"name":"param","in":"path","description":"Пороговое значение для размера головы","required":true,"schema":{"type":"integer","format":"int64"},"example":10}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DragonDTO"}}}}},"security":[{"bearerAuth":[]}]}},"/dragons/avgAge":{"get":{"tags":["Special"],"summary":"Средний возраст драконов","operationId":"getAvgAge","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"number","format":"double"}}}}},"security":[{"bearerAuth":[]}]}},"/deletePersonById/{id}":{"delete":{"tags":["Delete"],"summary":"Удалить персонажа","description":"Удаляет персонажа по идентификатору. Если используется как убийца — вернёт 400.","operationId":"deletePersonById","parameters":[{"name":"id","in":"path","description":"ID персонажа","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Удалено","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Нельзя удалить: есть зависимости","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/deleteLocationById/{id}":{"delete":{"tags":["Delete"],"summary":"Удалить локацию","description":"Удаляет локацию по идентификатору. Если используется — вернёт 400.","operationId":"deleteLocationById","parameters":[{"name":"id","in":"path","description":"ID локации","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Удалено","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Нельзя удалить: есть зависимости","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/deleteHeadById/{id}":{"delete":{"tags":["Delete"],"summary":"Удалить голову дракона","description":"Удаляет голову по идентификатору. Если используется — вернёт 400.","operationId":"deleteHeadById","parameters":[{"name":"id","in":"path","description":"ID головы","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Удалено","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Нельзя удалить: есть зависимости","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/deleteDragonById/{id}":{"delete":{"tags":["Dragon"],"summary":"Удалить дракона по ID","description":"Удаляет дракона, отправляет обновлённый список по WebSocket.","operationId":"deleteDragonById","parameters":[{"name":"id","in":"path","description":"ID дракона","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"201":{"description":"Удалено (сайд-эффект отправки списка)","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/deleteCoordinatesById/{id}":{"delete":{"tags":["Delete"],"summary":"Удалить координаты","description":"Удаляет координаты по идентификатору. Если координаты используются — вернёт 400.","operationId":"deleteCoordinatesById","parameters":[{"name":"id","in":"path","description":"ID координат","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Удалено","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Нельзя удалить: есть зависимости","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}},"/deleteCaveById/{id}":{"delete":{"tags":["Delete"],"summary":"Удалить пещеру","description":"Удаляет пещеру по идентификатору. Если пещера используется — вернёт 400.","operationId":"deleteCaveById","parameters":[{"name":"id","in":"path","description":"ID пещеры","required":true,"schema":{"type":"integer","format":"int64"},"example":1}],"responses":{"200":{"description":"Удалено","content":{"*/*":{"schema":{"type":"string"}}}},"400":{"description":"Нельзя удалить: есть зависимости","content":{"*/*":{"schema":{"type":"string"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"Coordinates":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"x":{"type":"number","format":"float"},"y":{"type":"number","format":"double"}},"required":["x","y"]},"Dragon":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string","minLength":1},"coordinates":{"$ref":"#/components/schemas/Coordinates"},"creationDate":{"type":"string","format":"date-time"},"cave":{"$ref":"#/components/schemas/DragonCave"},"killer":{"$ref":"#/components/schemas/Person"},"age":{"type":"integer","format":"int64"},"description":{"type":"string"},"wingspan":{"type":"number","format":"float"},"type":{"type":"string","enum":["WATER","UNDERGROUND","AIR","FIRE"]},"head":{"$ref":"#/components/schemas/DragonHead"}},"required":["cave","coordinates","description","name"]},"DragonCave":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"numberOfTreasures":{"type":"number","format":"float"}}},"DragonHead":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int64"},"eyesCount":{"type":"integer","format":"int32"},"toothCount":{"type":"integer","format":"int32"}}},"Location":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"x":{"type":"number","format":"float"},"y":{"type":"number","format":"float"},"z":{"type":"integer","format":"int32"},"name":{"type":"string"}},"required":["x","z"]},"Person":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string","minLength":1},"eyeColor":{"type":"string","enum":["RED","BLACK","YELLOW","WHITE","BROWN"]},"hairColor":{"type":"string","enum":["RED","BLACK","YELLOW","WHITE","BROWN"]},"location":{"$ref":"#/components/schemas/Location"},"passportID":{"type":"string","minLength":1},"nationality":{"type":"string","enum":["FRANCE","SPAIN","VATICAN","ITALY","NORTH_KOREA"]}},"required":["eyeColor","name","nationality","passportID"]},"CoordinatesDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"x":{"type":"number","format":"float"},"y":{"type":"number","format":"double"}},"required":["x","y"]},"DragonCaveDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"numberOfTreasures":{"type":"number","format":"float","minimum":0.0}}},"DragonDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string","minLength":1},"coordinates":{"$ref":"#/components/schemas/CoordinatesDTO"},"creationDate":{"type":"string","format":"date-time"},"cave":{"$ref":"#/components/schemas/DragonCaveDTO"},"killer":{"$ref":"#/components/schemas/PersonDTO"},"age":{"type":"integer","format":"int64"},"description":{"type":"string"},"wingspan":{"type":"number","format":"float"},"type":{"type":"string","enum":["WATER","UNDERGROUND","AIR","FIRE"]},"head":{"$ref":"#/components/schemas/DragonHeadDTO"}},"required":["cave","coordinates","creationDate","description","name"]},"DragonHeadDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"size":{"type":"integer","format":"int64","minimum":1},"eyesCount":{"type":"integer","format":"int32","minimum":1},"toothCount":{"type":"integer","format":"int32","minimum":1}},"required":["eyesCount","size","toothCount"]},"LocationDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"x":{"type":"number","format":"float"},"y":{"type":"number","format":"float"},"z":{"type":"integer","format":"int32"},"name":{"type":"string"}},"required":["x","y","z"]},"PersonDTO":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"eyeColor":{"type":"string","enum":["RED","BLACK","YELLOW","WHITE","BROWN"]},"hairColor":{"type":"string","enum":["RED","BLACK","YELLOW","WHITE","BROWN"]},"passportID":{"type":"string"},"nationality":{"type":"string","enum":["FRANCE","SPAIN","VATICAN","ITALY","NORTH_KOREA"]},"location":{"$ref":"#/components/schemas/LocationDTO"}}},"KillDragonDTO":{"type":"object","properties":{"dragonId":{"type":"integer","format":"int64"},"killerId":{"type":"integer","format":"int64"}},"required":["dragonId","killerId"]},"NewDragonResp":{"type":"object","description":"Запрос на аутентификацию","properties":{"name":{"type":"string","minLength":1},"coordinates":{"$ref":"#/components/schemas/Coordinates"},"cave":{"$ref":"#/components/schemas/DragonCave"},"killer":{"$ref":"#/components/schemas/Person"},"age":{"type":"integer","format":"int64"},"description":{"type":"string","minLength":1},"wingspan":{"type":"number","format":"float"},"type":{"type":"string","enum":["WATER","UNDERGROUND","AIR","FIRE"]},"head":{"$ref":"#/components/schemas/DragonHead"}},"required":["cave","coordinates","description","name","type"]},"SignUpRequest":{"type":"object","description":"Запрос на регистрацию","properties":{"username":{"type":"string","description":"Имя пользователя","example":"Jon","maxLength":20,"minLength":5},"password":{"type":"string","description":"Пароль","example":"my_1secret1_password","maxLength":255,"minLength":8}},"required":["password","username"]},"AuthRespForm":{"type":"object","properties":{"token":{"type":"string"},"username":{"type":"string"}}},"SignInRequest":{"type":"object","description":"Запрос на аутентификацию","properties":{"username":{"type":"string","description":"Имя пользователя","example":"Jon","maxLength":20,"minLength":5},"password":{"type":"string","description":"Пароль","example":"my_1secret1_password","maxLength":255,"minLength":8}},"required":["password","username"]}}}}